requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==3.2
django-crispy-forms==1.14.0
django-phone-field==1.8.1
django-phonenumber-field==6.4.0
idna==3.3
phonenumbers==8.12.54
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
